+++ src/main/java/net/openhft/chronicle/queue/internal/streaming/VanillaMapper.java
@@ -1,11 +1,11 @@
 package net.openhft.chronicle.queue.internal.streaming;
 
-import net.openhft.chronicle.queue.AppenderListener;
+import net.openhft.chronicle.queue.incubator.streaming.Accumulation;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.function.Function;
 
-final class VanillaMapper<T, R> implements AppenderListener.Accumulation.MapperTo<R> {
+final class VanillaMapper<T, R> implements Accumulation.MapperTo<R> {
 
     private final T delegate;
     private final Function<? super T, ? extends R> mapper;
