+++ src/test/java/net/openhft/chronicle/queue/incubator/streaming/StreamsTest.java
@@ -1,4 +1,4 @@
-package net.openhft.chronicle.queue.streaming;
+package net.openhft.chronicle.queue.incubator.streaming;
 
 import net.openhft.chronicle.core.io.IOTools;
 import net.openhft.chronicle.core.time.SetTimeProvider;
